% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedrocktrader-package.R
\docType{package}
\name{bedrocktrader}
\alias{bedrocktrader}
\alias{bedrocktrader-package}
\title{bedrocktrader: Retrieve Minecraft Bedrock Villager Trade Data}
\description{
\code{bedrocktrader} retrieves official vanilla Minecraft Bedrock villager trade
data from stable Mojang releases and records exact source provenance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rentosaijo.github.io/bedrocktrader/}
  \item \url{https://github.com/RentoSaijo/bedrocktrader}
  \item Report bugs at \url{https://github.com/RentoSaijo/bedrocktrader/issues}
}

}
\author{
\strong{Maintainer}: Rento Saijo \email{rentosaijo0527@gmail.com} (\href{https://orcid.org/0009-0008-4919-7349}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Rento Saijo \email{rentosaijo0527@gmail.com} (\href{https://orcid.org/0009-0008-4919-7349}{ORCID}) [copyright holder]
}

}
